refactor(in-game): runGame.js 삭제
